requirements.txt
setuptools~=57.0.0
pip~=21.1.2
pygame~=2.0.1
pygame-menu~=4.1.0
psycopg2~=2.8.6
python-dotenv~=0.17.1